moved source into rrt subdir
this makes it easier to use this library inside of other projects because
all #includes are relative to the base directory

for example, before in our code we did #include "StateSpace.hpp"
we now do #include <rrt/StateSpace.hpp>
this essentially "namespaces" all included files, reducing collisions with external projects
that include this one
