+++ src/rrt-viewer/RRTWidget.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 #include <QtWidgets>
-#include <BiRRT.hpp>
-#include <2dplane/GridStateSpace.hpp>
+#include <rrt/BiRRT.hpp>
+#include <rrt/2dplane/GridStateSpace.hpp>
 #include <Eigen/Dense>
