+++ src/rrt/2dplane/2dplane.cpp
@@ -1,5 +1,5 @@
 
-#include "2dplane.hpp"
+#include <rrt/2dplane/2dplane.hpp>
 #include <functional>
 #include <math.h>
 
@@ -8,7 +8,7 @@ using namespace RRT;
 using namespace std;
 
 
-Tree<Vector2f> *TreeFor2dPlane(shared_ptr<StateSpace<Eigen::Vector2f>> stateSpace,
+Tree<Vector2f> *RRT::TreeFor2dPlane(shared_ptr<StateSpace<Eigen::Vector2f>> stateSpace,
     Vector2f goal,
     float step) {
     Tree<Vector2f> *rrt = new Tree<Vector2f>(stateSpace);
