+++ src/rrt/2dplane/2dplane.hpp
@@ -1,6 +1,6 @@
 
-#include <Tree.hpp>
-#include <StateSpace.hpp>
+#include <rrt/Tree.hpp>
+#include <rrt/StateSpace.hpp>
 #include <Eigen/Dense>
 
 namespace RRT {
