+++ src/rrt/2dplane/PlaneStateSpace.cpp
@@ -1,4 +1,4 @@
-#include "PlaneStateSpace.hpp"
+#include <rrt/2dplane/PlaneStateSpace.hpp>
 
 using namespace Eigen;
