+++ src/rrt/2dplane/PlaneStateSpace.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <StateSpace.hpp>
+#include <rrt/StateSpace.hpp>
 #include <Eigen/Dense>
 
 namespace RRT {
