+++ src/rrt/BiRRT.hpp
@@ -1,4 +1,4 @@
-#include "Tree.hpp"
+#include <rrt/Tree.hpp>
 #include <limits.h>
