+++ src/rrt/Tree.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <StateSpace.hpp>
+#include <rrt/StateSpace.hpp>
 #include <list>
 #include <vector>
 #include <memory>
