+++ src/rrt/TreeTest.cpp
@@ -1,7 +1,7 @@
 #include "gtest/gtest.h"
-#include <Tree.hpp>
-#include <2dplane/2dplane.hpp>
-#include <2dplane/GridStateSpace.hpp>
+#include <rrt/Tree.hpp>
+#include <rrt/2dplane/2dplane.hpp>
+#include <rrt/2dplane/GridStateSpace.hpp>
 #include <vector>
 
 using namespace Eigen;
