+++ src/rrt/planning/Path.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "StateSpace.hpp"
+#include <rrt/StateSpace.hpp>
 #include <vector>
